% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACTncdGer.R
\docType{package}
\name{IMPACTncdGer}
\alias{IMPACTncdGer-package}
\alias{IMPACTncdGer}
\title{IMPACTncdGer: A package with functions for the IMPACTncdGer model.}
\description{
Description of your package.
}
\author{
Chris Kypridemos
}
